Remove the `eventlisteners` module.
The `eventlisteners` snabbdom module allows you to add event listeners
when creating a virtual node via the `h` method. You could for
example do this in the `toVNode` method, but not in the `toHTML` method.

This way of registering event listeners is in contrast to Backbone's
declarative way of registering event, which is more the "Backbone way".

Since in the Backbone.VDOMView we undelegate and re-delegate all
declarative events at each `render` (via `setElement`), we already have a
powerful way to register event listeners.